refactor(hooks): reorganize theme and monitor hooks into separate directories
Restructure theme and monitor-related hooks into dedicated directories for better maintainability and clarity. This includes moving `useThemeState`, `useThemeSelection`, `useThemeApplication`, `useMonitorSelection`, and `useMonitorThemeSync` into `theme` and `monitor` subdirectories respectively. Remove deprecated `ThemeContext.tsx` file as its functionality is now handled by `useThemeManager.tsx`.
