--- src/components/ThemeContext.tsx
@@ -1,6 +1,0 @@
-/**
- * 此文件仅作为兼容层，实际实现已移至 hooks/useThemeManager.tsx
- * 为避免修改引用关系，保留此文件并重定向到新实现
- */
-
-export { useThemeSelector } from "~/hooks/useThemeManager";
